review comment changes for code optimization
